Lexical processor refactored. Unit tests added for lexical processor Less than & greater than added to MQL
