+++ tests/unit-tests/orm_test.ts
@@ -4,6 +4,9 @@ import { mockDecorator, Orange, Test, DenoAsserts } from "../mod.ts";
 import { Types } from "../../orm-core/sql/types.ts";
 import { MandarineRepository } from "../../orm-core/repository/mandarineRepository.ts";
 import { RepositoryComponent } from "../../main-core/components/repository-component/repositoryComponent.ts";
+import { lexicalProcessor } from "../../orm-core/core/lexicalProcessor.ts";
+import { PostgreSQLDialect } from "../../orm-core/dialect/postgreSQLDialect.ts";
+import { Mandarine } from "../../main-core/Mandarine.ns.ts";
 
 @mockDecorator()
 class MyTable {
@@ -145,6 +148,79 @@ export class ORMTests {
         DenoAsserts.assert(typeof handler.findAll === 'function');
         DenoAsserts.assert(typeof handler.deleteAll === 'function');
         DenoAsserts.assert(typeof handler.countAll === 'function');
+    }
+
+    @Test({
+        name: "Lexical processor",
+        description: "Should create SQL queries"
+    })
+    public useLexicalProcessor() {
+
+        // @ts-ignore
+        const fakeRepositoryProxy: Mandarine.ORM.RepositoryProxy = {
+            SUPPORTED_KEYWORDS: ["and", "or", "isnotnull", "isnull", "isempty", "isnotempty", "startingwith", "endswith", "like", "greaterthan", "lessthan"]
+        };
+
+        const fakeTableMetadata = {
+            name: "users",
+            schema: "public"
+        };
+
+        // @ts-ignore
+        const fakeEntity: Mandarine.ORM.Entity.Table = {
+            columns: [
+                {
+                    name: "id"
+                },
+                {
+                    name: "firstname"
+                },
+                {
+                    name: "lastname"
+                },
+                {
+                    name: "age"
+                },
+                {
+                    name: "posts"
+                },
+                {
+                    name: "country"
+                }
+            ]
+        }
+
+        const dialect = new PostgreSQLDialect();
+
+        const countByCountry = lexicalProcessor(fakeRepositoryProxy, "countByCountry", "countBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(countByCountry, "SELECT COUNT(*) FROM public.users WHERE country = $1");
+
+        const findByCountry = lexicalProcessor(fakeRepositoryProxy, "findByCountry", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByCountry, "SELECT * FROM public.users WHERE country = $1");
+
+        const findByFirstnameAndCountry = lexicalProcessor(fakeRepositoryProxy, "findByFirstnameAndCountry", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByFirstnameAndCountry, "SELECT * FROM public.users WHERE firstname = $1 AND country = $2");
+
+        const findByCountryIsNotNull = lexicalProcessor(fakeRepositoryProxy, "findByCountryIsNotNull", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByCountryIsNotNull, "SELECT * FROM public.users WHERE country IS NOT NULL");
+
+        const findByCountryIsNull = lexicalProcessor(fakeRepositoryProxy, "findByCountryIsNull", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByCountryIsNull, "SELECT * FROM public.users WHERE country IS NULL");
+
+        const findByFirstnameIsNotEmptyAndCountryIsEmpty = lexicalProcessor(fakeRepositoryProxy, "findByFirstnameIsNotEmptyAndCountryIsEmpty", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByFirstnameIsNotEmptyAndCountryIsEmpty, "SELECT * FROM public.users WHERE firstname <> '' AND country = ''");
+
+        const findByLastnameStartingWith = lexicalProcessor(fakeRepositoryProxy, "findByLastnameStartingWith", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByLastnameStartingWith, "SELECT * FROM public.users WHERE lastname LIKE '' || $1 || '%'");
+
+        const findByLastnameAndCountryLikeAndFirstnameEndsWith = lexicalProcessor(fakeRepositoryProxy, "findByLastnameAndCountryLikeAndFirstnameEndsWith", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByLastnameAndCountryLikeAndFirstnameEndsWith, "SELECT * FROM public.users WHERE lastname = $1 AND country LIKE '%' || $2 || '%' AND firstname LIKE '%' || $3 || ''");
+
+        const findByFirstnameAndCountryAndLastnameEndsWithAndFirstnameIsNotNullOrLastnameIsNullAndCountryLike = lexicalProcessor(fakeRepositoryProxy, "findByFirstnameAndCountryAndLastnameEndsWithAndFirstnameIsNotNullOrLastnameIsNullAndCountryLike", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByFirstnameAndCountryAndLastnameEndsWithAndFirstnameIsNotNullOrLastnameIsNullAndCountryLike, "SELECT * FROM public.users WHERE firstname = $1 AND country = $2 AND lastname LIKE '%' || $3 || '' AND firstname IS NOT NULL OR lastname IS NULL AND country LIKE '%' || $4 || '%'");
+    
+        const findByFirstnameAndAgeGreaterThanAndPostsLessThanOrPostsGreaterThan = lexicalProcessor(fakeRepositoryProxy, "findByFirstnameAndAgeGreaterThanAndPostsLessThanOrPostsGreaterThan", "findBy", fakeTableMetadata, fakeEntity, dialect);
+        DenoAsserts.assertEquals(findByFirstnameAndAgeGreaterThanAndPostsLessThanOrPostsGreaterThan, "SELECT * FROM public.users WHERE firstname = $1 AND age > $2 AND posts < $3 OR posts > $4")
     }
 
 }
